@@ -5,7 +5,7 @@ import 'package:flutter/widgets.dart';
 /// Custom Wrap
 class CustomWrap extends MultiChildRenderObjectWidget {
   CustomWrap({
-    Key key,
+    Key? key,
     this.column,
     this.symmetry,
     this.direction = Axis.horizontal,
@@ -19,9 +19,9 @@ class CustomWrap extends MultiChildRenderObjectWidget {
     List<Widget> children = const <Widget>[],
   }) : super(key: key, children: children);
 
-  final int column;
+  final int? column;
 
-  final bool symmetry;
+  final bool? symmetry;
 
   final Axis direction;
 
@@ -35,7 +35,7 @@ class CustomWrap extends MultiChildRenderObjectWidget {
 
   final WrapCrossAlignment crossAxisAlignment;
 
-  final TextDirection textDirection;
+  final TextDirection? textDirection;
 
   final VerticalDirection verticalDirection;
 
@@ -93,16 +93,16 @@ class CustomRenderWrap extends RenderBox
         ContainerRenderObjectMixin<RenderBox, WrapParentData>,
         RenderBoxContainerDefaultsMixin<RenderBox, WrapParentData> {
   CustomRenderWrap({
-    List<RenderBox> children,
-    int column,
-    bool symmetry,
+    List<RenderBox>? children,
+    int? column,
+    bool? symmetry,
     Axis direction = Axis.horizontal,
     WrapAlignment alignment = WrapAlignment.start,
     double spacing = 0.0,
     WrapAlignment runAlignment = WrapAlignment.start,
     double runSpacing = 0.0,
     WrapCrossAlignment crossAxisAlignment = WrapCrossAlignment.start,
-    TextDirection textDirection,
+    TextDirection? textDirection,
     VerticalDirection verticalDirection = VerticalDirection.down,
   })  : assert(direction != null),
         assert(alignment != null),
@@ -123,17 +123,17 @@ class CustomRenderWrap extends RenderBox
     addAll(children);
   }
 
-  int get column => _column;
-  int _column;
-  set column(int value) {
+  int? get column => _column;
+  int? _column;
+  set column(int? value) {
     if (column == value) return;
     _column = value;
     markNeedsLayout();
   }
 
-  bool get symmetry => _symmetry;
-  bool _symmetry;
-  set symmetry(bool value) {
+  bool? get symmetry => _symmetry;
+  bool? _symmetry;
+  set symmetry(bool? value) {
     if (symmetry == value) return;
     _symmetry = value;
     markNeedsLayout();
@@ -193,9 +193,9 @@ class CustomRenderWrap extends RenderBox
     markNeedsLayout();
   }
 
-  TextDirection get textDirection => _textDirection;
-  TextDirection _textDirection;
-  set textDirection(TextDirection value) {
+  TextDirection? get textDirection => _textDirection;
+  TextDirection? _textDirection;
+  set textDirection(TextDirection? value) {
     if (_textDirection != value) {
       _textDirection = value;
       markNeedsLayout();
@@ -283,7 +283,7 @@ class CustomRenderWrap extends RenderBox
     double runWidth = 0.0;
     double runHeight = 0.0;
     int childCount = 0;
-    RenderBox child = firstChild;
+    RenderBox? child = firstChild;
     while (child != null) {
       final double childWidth = child.getMaxIntrinsicWidth(double.infinity);
       final double childHeight = child.getMaxIntrinsicHeight(childWidth);
@@ -312,7 +312,7 @@ class CustomRenderWrap extends RenderBox
     double runHeight = 0.0;
     double runWidth = 0.0;
     int childCount = 0;
-    RenderBox child = firstChild;
+    RenderBox? child = firstChild;
     while (child != null) {
       final double childHeight = child.getMaxIntrinsicHeight(double.infinity);
       final double childWidth = child.getMaxIntrinsicWidth(childHeight);
@@ -339,16 +339,16 @@ class CustomRenderWrap extends RenderBox
     switch (direction) {
       case Axis.horizontal:
         double width = 0.0;
-        RenderBox child = firstChild;
+        RenderBox? child = firstChild;
         while (child != null) {
           width = math.max(width, child.getMinIntrinsicWidth(double.infinity));
           child = childAfter(child);
         }
         return width;
       case Axis.vertical:
+      default:
         return _computeIntrinsicWidthForHeight(height);
     }
-    return null;
   }
 
   @override
@@ -356,16 +356,16 @@ class CustomRenderWrap extends RenderBox
     switch (direction) {
       case Axis.horizontal:
         double width = 0.0;
-        RenderBox child = firstChild;
+        RenderBox? child = firstChild;
         while (child != null) {
           width += child.getMaxIntrinsicWidth(double.infinity);
           child = childAfter(child);
         }
         return width;
       case Axis.vertical:
+      default:
         return _computeIntrinsicWidthForHeight(height);
     }
-    return null;
   }
 
   @override
@@ -374,8 +374,9 @@ class CustomRenderWrap extends RenderBox
       case Axis.horizontal:
         return _computeIntrinsicHeightForWidth(width);
       case Axis.vertical:
+      default:
         double height = 0.0;
-        RenderBox child = firstChild;
+        RenderBox? child = firstChild;
         while (child != null) {
           height =
               math.max(height, child.getMinIntrinsicHeight(double.infinity));
@@ -383,7 +384,6 @@ class CustomRenderWrap extends RenderBox
         }
         return height;
     }
-    return null;
   }
 
   @override
@@ -392,19 +392,19 @@ class CustomRenderWrap extends RenderBox
       case Axis.horizontal:
         return _computeIntrinsicHeightForWidth(width);
       case Axis.vertical:
+      default:
         double height = 0.0;
-        RenderBox child = firstChild;
+        RenderBox? child = firstChild;
         while (child != null) {
           height += child.getMaxIntrinsicHeight(double.infinity);
           child = childAfter(child);
         }
         return height;
     }
-    return null;
   }
 
   @override
-  double computeDistanceToActualBaseline(TextBaseline baseline) {
+  double? computeDistanceToActualBaseline(TextBaseline baseline) {
     return defaultComputeDistanceToHighestActualBaseline(baseline);
   }
 
@@ -458,12 +458,12 @@ class CustomRenderWrap extends RenderBox
   void performLayout() {
     assert(_debugHasNecessaryDirections);
     _hasVisualOverflow = false;
-    RenderBox child = firstChild;
+    RenderBox? child = firstChild;
     if (child == null) {
       size = constraints.smallest;
       return;
     }
-    BoxConstraints childConstraints;
+    BoxConstraints? childConstraints;
     double mainAxisLimit = 0.0;
     bool flipMainAxis = false;
     bool flipCrossAxis = false;
@@ -516,7 +516,7 @@ class CustomRenderWrap extends RenderBox
       if (childCount > 0) runMainAxisExtent += spacing;
       runCrossAxisExtent = math.max(runCrossAxisExtent, childCrossAxisExtent);
       childCount += 1;
-      final WrapParentData childParentData = child.parentData;
+      final WrapParentData childParentData = child.parentData as WrapParentData;
       childParentData._runIndex = runMetrics.length;
       child = childParentData.nextSibling;
     }
@@ -597,11 +597,11 @@ class CustomRenderWrap extends RenderBox
       //print(symmetry);
       switch (alignment) {
         case WrapAlignment.start:
-          if (symmetry) childLeadingSpace = spacing / 2;
+          if (symmetry!) childLeadingSpace = spacing / 2;
           break;
         case WrapAlignment.end:
           childLeadingSpace = mainAxisFreeSpace;
-          if (symmetry) childLeadingSpace = mainAxisFreeSpace - spacing / 2;
+          if (symmetry!) childLeadingSpace = mainAxisFreeSpace - spacing / 2;
           break;
         case WrapAlignment.center:
           childLeadingSpace = mainAxisFreeSpace / 2.0;
@@ -628,7 +628,8 @@ class CustomRenderWrap extends RenderBox
       if (flipCrossAxis) crossAxisOffset -= runCrossAxisExtent;
 
       while (child != null) {
-        final WrapParentData childParentData = child.parentData;
+        final WrapParentData childParentData =
+            child.parentData as WrapParentData;
         if (childParentData._runIndex != i) break;
         final double childMainAxisExtent = _getMainAxisExtent(child);
         final double childCrossAxisExtent = _getCrossAxisExtent(child);
@@ -652,8 +653,9 @@ class CustomRenderWrap extends RenderBox
   }
 
   @override
-  bool hitTestChildren(HitTestResult result, {Offset position}) {
-    return defaultHitTestChildren(result, position: position);
+  bool hitTestChildren(HitTestResult result, {required Offset position}) {
+    return defaultHitTestChildren(result as BoxHitTestResult,
+        position: position);
   }
 
   @override
